requirements.txt
fastapi
pymysql
sqlmodel
pydantic
uvicorn
python-dotenv
typing
python-jose

